Added ImageMagick functionality (#2426)
* Added ImageMagick functionality

Fixed forgotten openssl migration to openssl3
Minor code cleanup

* Fixed forgotten code cleaning

* Fixed: Extension check before conversion

Added: Mogrify command method to ImageMagick
Fixed: jxl conversion (using mogrify method)
Changed: How FQDN functions to cater to wider range of usage
Fixed: Checking of files before download to include converted files too

* Changed: ImageMagick hint for better visibility

* Fixed: ImageMagick hint text

Fixed: FQDN files/folders not being deleted
Changed: Check if file exists to include archives

* Fixed: Deletion of FQDN files/folders properly this time

+++ baseunits/uPacker.pas
@@ -123,8 +123,7 @@ function TPacker.Do7Zip: Boolean;
 var
   p: TProcess;
   exit_status, i: Integer;
-  s, sout, serr, fFileName,
-  fUniqueTimestampName: string;
+  s, sout, serr, fFileName: String;
 begin
   Result := False;
   p := TProcess.Create(nil);
@@ -138,36 +137,7 @@ begin
       end;
     end;
 
-    fUniqueTimestampName := StringReplace(ExtractFileName(FSavedFileName), ExtractFileExt(FSavedFileName), '', [rfReplaceAll])+ '_' + FormatDateTime('yyyy-mm-dd_hh-nn-ss', Now);
-    fFileName := 'FQDNList_' + fUniqueTimestampName + '.txt';
-    with TStringList.Create do
-    try
-      for i := 0 to FFileList.Count - 1 do
-      begin
-        FFileList[i] := MainForm.CheckLongNamePaths(FFileList[i]);
-        Add(FFileList[i]);
-      end;
-
-      try
-        SaveToFile(fFileName);
-      except
-        on E: EFCreateError do
-        begin
-          // If first attempt fails, try the second path
-          try
-            fFileName := AppendPathDelim(ExtractFilePath(FSavedFileName)) + fFileName;
-            SaveToFile(fFileName);
-          except
-            on E2: Exception do
-            begin
-              MainForm.ExceptionHandler(Self, E2);
-            end;
-          end;
-        end;
-      end;
-    finally
-      Free;
-    end;
+    fFileName := CreateFQDNList(Self, FSavedFileName, FFileList);
 
     p.Executable := CURRENT_ZIP_EXE;
     with p.Parameters do begin
@@ -191,7 +161,11 @@ begin
     p.ShowWindow := swoHIDE;
     p.RunCommandLoop(sout, serr, exit_status);
     Result := exit_status = 0;
-    DeleteFile(fFileName);
+
+    if FileExists(fFileName) then
+    begin
+      DeleteFile(fFileName);
+    end;
 
     if not Result then
     begin
@@ -320,11 +294,6 @@ begin
       end;
   end;
 
-  if FFileList.Count = 0 then
-  begin
-    Exit;
-  end;
-
   FFileList.CustomSort(NaturalCustomSort);
   case Format of
     pfZIP: FExt := '.zip';
@@ -340,6 +309,16 @@ begin
   else
   begin
     FSavedFileName := TrimAndExpandFilename(Path) + FExt;
+  end;
+
+  if FFileList.Count = 0 then
+  begin
+    if FileExists(FSavedFileName) then
+    begin
+      Exit(True);
+    end;
+
+    Exit;
   end;
 
   if FileExists(FSavedFileName) then
